Project Code Initial cut
The dev container and files included are basically just copies of homework 4. I made some updates to the python code to account of some of the differences we have. It does not run as is since Im still trying to work out the dynamics equations for a three link arm as opposed to a two link arm like the homework.